Refactored and splitted crypto module into several modules
The structured directoy layout is the following:

* ciphers are now located in sys/crypto/chipers/*
* hashes are now located in sys/crypto/hashes/*

Each cipher (AES, 3DES, RC5, SKIPJACK, TWOFISH) / hash (sha256)
can be used with USEMODULE-Macro.
