+++ sys/crypto/ciphers/3des/Makefile
@@ -1,0 +1,3 @@
+MODULE = 3des
+
+include $(RIOTBASE)/Makefile.base
