+++ sys/crypto/ciphers/Makefile
@@ -1,0 +1,3 @@
+MODULE = crypto_ciphers
+
+include $(RIOTBASE)/Makefile.base
