+++ sys/crypto/ciphers/aes/Makefile
@@ -1,0 +1,3 @@
+MODULE = aes
+
+include $(RIOTBASE)/Makefile.base
