+++ sys/crypto/ciphers/aes/aes.c
@@ -33,7 +33,7 @@
 #include <string.h>
 #include <stdlib.h>
 #include <stdint.h>
-#include "crypto/aes.h"
+#include "crypto/ciphers/aes.h"
 #include "crypto/ciphers.h"
 
 /**
