+++ sys/crypto/ciphers/rc5/Makefile
@@ -1,0 +1,3 @@
+MODULE = rc5
+
+include $(RIOTBASE)/Makefile.base
