+++ sys/crypto/ciphers/skipjack/Makefile
@@ -1,0 +1,3 @@
+MODULE = skipjack
+
+include $(RIOTBASE)/Makefile.base
