+++ sys/crypto/ciphers/twofish/Makefile
@@ -1,0 +1,3 @@
+MODULE = twofish
+
+include $(RIOTBASE)/Makefile.base
