+++ sys/crypto/ciphers/twofish/twofish.c
@@ -26,7 +26,7 @@
 #include <string.h>
 #include <stdlib.h>
 #include <stdint.h>
-#include "crypto/twofish.h"
+#include "crypto/ciphers/twofish.h"
 #include "crypto/ciphers.h"
