+++ sys/crypto/hashes/sha256/Makefile
@@ -1,1 +1,3 @@
+MODULE = sha256
+
 include $(RIOTBASE)/Makefile.base
