+++ tests/test_sha256/Makefile
@@ -1,7 +1,7 @@
 APPLICATION = test_sha256
 include ../Makefile.tests_common
 
-USEMODULE += crypto
+USEMODULE += sha256
 
 DISABLE_MODULE += auto_init
