+++ tests/test_sha256/main.c
@@ -22,7 +22,7 @@
 #include <stdio.h>
 #include <string.h>
 
-#include "crypto/sha256.h"
+#include "crypto/hashes/sha256.h"
 
 unsigned char hash[SHA256_DIGEST_LENGTH];
